Fix progress bar usage
Added parameter is_editable to progress_bar decorator which is set to
False by default. If set to False, no additional operation_status
parameter is needed in the decorated function and progress bar simply
shows default operation status (or initial_status which is also
possible to newly set). If set to True, operation_status parameter is
needed in decorated function and via this parameter the progress bar
can be paused and status updated.